+++ app/components/Sidebar/components/DocumentLink.tsx
@@ -25,7 +25,7 @@ import {
   useDropToReorderDocument,
   useDropToReparentDocument,
 } from "../hooks/useDragAndDrop";
-import { useSidebarExpansion } from "../hooks/useSidebarExpansion";
+import { useSidebarExpansion } from "./SidebarExpansionContext";
 import DocumentRow from "./DocumentRow";
 import DropCursor from "./DropCursor";
 import Folder from "./Folder";
